// src/types/datapoint.ts
export interface DataBlock {
  /**
   * A human-readable summary of this data block.
   */
  summary: string;

  /**
   * A machine-readable text summary of this data block.
   */
  icon: WeatherIcon;

  /**
   * An array of data points, ordered by time, which together describe the weather conditions
   * at the requested location over time.
   */
  data: DataPoint[];
}
// ...
export interface CurrentlyDataPoint extends DataPoint {
  summary: string;
  icon: WeatherIcon;
  nearestStormDistance: number;
  precipIntensity: number;
  precipIntensityError: number;
  precipProbability: number;
  precipType: "rain";
  temperature: number;
  apparentTemperature: number;
  dewPoint: number;
  humidity: number;
  pressure: number;
  windSpeed: number;
  windGust: number;
  cloudCover: number;
  uvIndex: number;
  visibility: number;
  ozone: number;
}
// ...
export interface HourlyDataBlock extends DataBlock {
  data: HourlyDataPoint[];
}

/**
 * A DataPoint that is returned from the Hourly data block.
 *
 * @see DataPoint
 */
export interface HourlyDataPoint extends DataPoint {
  summary: string;
  icon: WeatherIcon;
  precipIntensity: number;
  precipProbability: number;
  temperature: number;
  apparentTemperature: number;
  dewPoint: number;
  humidity: number;
  pressure: number;
  windSpeed: number;
  windGust: number;
  cloudCover: number;
  uvIndex: number;
  visibility: number;
  ozone: number;
}

/**
 * A DataBlock that is returned from the Daily data block.
 *
 * @see DataBlock
 */
export interface DailyDataBlock extends DataBlock {
  data: DailyDataPoint[];
}

/**
 * A DataPoint that is returned from the Daily data block.
 *
 * @see DataPoint
 */
export interface DailyDataPoint extends DataPoint {
  summary: string;
  icon: WeatherIcon;
  sunriseTime: number;
  sunsetTime: number;
  moonPhase: number;
  precipIntensity: number;
  precipIntensityMax: number;
  precipIntensityMaxTime: number;
  precipProbability: number;
  temperatureHigh: number;
  temperatureHighTime: number;
  temperatureLow: number;
  temperatureLowTime: number;
  apparentTemperatureHigh: number;
  apparentTemperatureHighTime: number;
  apparentTemperatureLow: number;
  apparentTemperatureLowTime: number;
  dewPoint: number;
  humidity: number;
  pressure: number;
  windSpeed: number;
  windGust: number;
  windGustTime: number;
  cloudCover: number;
  uvIndex: number;
  uvIndexTime: number;
  visibility: number;
  ozone: number;
  temperatureMin: number;
  temperatureMinTime: number;
  temperatureMax: number;
  temperatureMaxTime: number;
  apparentTemperatureMin: number;
  apparentTemperatureMinTime: number;
  apparentTemperatureMax: number;
  apparentTemperatureMaxTime: number;
}